pyproject.toml
[tool.poetry]
name = "masp"
version = "0.1.0"
description = "Not the Museu de Arte de São Paulo!"
authors = ["Pierre-Alexandre <beatsontapes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"


[tool.poetry.group.dev.dependencies]
unstructured = "^0.12.6"
flask = "^3.0.2"
flask-cors = "^4.0.0"
requests = "^2.31.0"
pytest = "^8.1.1"
langchain = "^0.1.13"
langchain-community = "^0.0.29"
langchainhub = "^0.1.15"
langchain-openai = "^0.1.0"
chromadb = "^0.4.24"
bs4 = "^0.0.2"
beautifulsoup4 = "^4.12.3"
ipykernel = "^6.29.3"
selenium = "^4.18.1"
sentence-transformers = "^2.6.0"
openai = "^1.14.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
